requirements.txt
torch
transformers
easyocr
streamlit
pillow
